+++ packages/codemod-rewrite-module-specifiers-to-full-paths/src/bin/codemod-dev.ts
@@ -1,2 +1,2 @@
-import '@pkerschbaum/runtime-extensions-node';
+import '@patricktree/runtime-extensions-node';
 import '#pkg/bin/codemod.js';