+++ packages/codemod-rewrite-module-specifiers-to-full-paths/src/load-typescript-program.ts
@@ -2,7 +2,7 @@ import path from 'node:path';
 import invariant from 'tiny-invariant';
 import ts from 'typescript';
 
-import { fsUtils } from '@pkerschbaum/commons-node/utils/fs';
+import { fsUtils } from '@patricktree/commons-node/utils/fs';
 
 export type TypeScriptProgram = {
   compilerOptions: ts.CompilerOptions;